bin/rfid-player: multiple media files per rfid
If you put several audio files in an RFID directory, they will all be
played one after the other.  RFID cards can be used to navigate between
the files. Navigation cards can be registered by creating a "NEXT" file
or a "PREV" file in their rfid directory. The "NEXT" card is used to
jump to the next track, "PREV" is used to jump to the previous one.

After a track in the child process is finished, a USR1 signal is sent
to the main process to start the next track.